Since it wouldn't update the master branch I have gone and pushed the final working version we have to this branch.
Commented out the pause via p button system since issue is that it fires so frequently that it passes p as pressed to be continually true and thus never stops toggling the game state. This is the current version we will be presenting since it has the cores met and works.